booth-app/requirements.txt
flask==3.1.1
Pillow==11.2.1
qrcode==7.3.1
python-dotenv==0.19.1
requests==2.26.0
boto3==1.34.81